python-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2==3.0.1
scikit-learn==1.3.2
numpy==1.26.2
nltk==3.8.1
requests==2.31.0
python-dotenv==1.0.0
hypothesis==6.92.1
psycopg2-binary==2.9.9
reportlab==4.0.7
